// src/app/models/proveedor.model.ts
export interface IProveedor {
  codEstadoDocumentos: true;
  codEstadoInspector: number;
  codUser: string;
  codUserUpdate: string;
  codigoDepartamento: string;
  fechaRegistro: Date;
  fechaRegistroUpdate: Date;
  id: number;
  info: string;
  infoUpdate: string;
  nombreCiudad: string;
  nombreDepartamento: string;
  prvCodBanco: number;
  prvCodCiudad: string;
  prvCodTipoCuenta: number;
  prvCodTipoProveeedor: number;
  prvCodTraza: number;
  prvContacto: string;
  prvCpaCodCondicionesPago: number;
  prvCpaContadoCual: string;
  prvCpaCual: string;
  prvDeclaramientoInhabilidadesInteres: boolean;
  prvDireccion: string;
  prvDtllesBanNroCuenta: string;
  prvExperienciaSector: boolean;
  prvFechaEnvio: Date;
  prvMail: string;
  prvNit: string;
  prvNombreProveedor: string;
  prvPoliticaTratamientoDatosPersonales: Boolean;
  prvProveeedor: number;
  prvRteLegalApellido: string;
  prvRteLegalCodCiudad: string;
  prvRteLegalEmail: string;
  prvRteLegalIdentificacion: string;
  prvRteLegalNombre: string;
  prvRteLegalTelefonoMovil: string;
  prvTelefono: string;
  prvTipoProveedorCual: string;
  prvUrlPdf: string;
  prvValidationNumber: number;
  usrEstado: number;
  valGestor: boolean;
}